refactor(settings): split integrations card components
