+++ apps/emdash-desktop/src/renderer/features/settings/components/GitHubAccountsSection.tsx
@@ -19,9 +19,56 @@ import type { GitHubAccountSummary } from '@shared/github';
 export function GitHubAccountsSection() {
   const { data: accounts = [], isLoading } = useGitHubAccounts();
   const sortedAccounts = sortGitHubAccountsByDefault(accounts);
+  const showConnectGitHub = useShowModal('githubConnectModal');
+
+  return (
+    <section className="space-y-3">
+      <div className="mb-3 flex items-center justify-between gap-3">
+        <div className="min-w-0">
+          <p className="text-sm text-foreground">GitHub</p>
+        </div>
+        <TooltipProvider delay={150}>
+          <Tooltip>
+            <TooltipTrigger>
+              <Button
+                type="button"
+                variant="ghost"
+                size="icon-sm"
+                onClick={() => showConnectGitHub({})}
+                aria-label="Add GitHub account"
+              >
+                <Plus className="h-3.5 w-3.5" />
+              </Button>
+            </TooltipTrigger>
+            <TooltipContent side="top">Add GitHub account</TooltipContent>
+          </Tooltip>
+        </TooltipProvider>
+      </div>
+
+      {isLoading ? (
+        <div className="text-muted-foreground flex items-center gap-2 py-3 text-sm">
+          <Loader2 className="h-4 w-4 animate-spin" />
+          Loading GitHub accounts...
+        </div>
+      ) : sortedAccounts.length === 0 ? (
+        <div className="flex items-center gap-3 rounded-lg border border-dashed border-border/70 p-3">
+          <div className="bg-muted/50 flex h-9 w-9 shrink-0 items-center justify-center rounded-md">
+            <Github className="text-muted-foreground h-4 w-4" />
+          </div>
+          <p className="text-muted-foreground min-w-0 flex-1 text-sm">
+            No GitHub accounts are connected.
+          </p>
+        </div>
+      ) : (
+        <GitHubAccountRows accounts={sortedAccounts} />
+      )}
+    </section>
+  );
+}
+
+export function GitHubAccountRows({ accounts }: { accounts: GitHubAccountSummary[] }) {
   const setDefaultMutation = useSetDefaultGitHubAccount();
   const removeMutation = useRemoveGitHubAccount();
-  const showConnectGitHub = useShowModal('githubConnectModal');
   const showConfirmRemove = useShowModal('confirmActionModal');
   const { toast } = useToast();
 
@@ -66,6 +113,7 @@ export function GitHubAccountsSection() {
         description = `This account is used by ${projectLabel}. Removing it will disable GitHub features for those projects until another GitHub account is assigned.`;
       }
     } catch {}
+
     showConfirmRemove({
       title: `Remove @${account.login}?`,
       description,
@@ -75,58 +123,20 @@ export function GitHubAccountsSection() {
   };
 
   return (
-    <section className="space-y-3">
-      <div className="mb-3 flex items-center justify-between gap-3">
-        <div className="min-w-0">
-          <p className="text-sm text-foreground">GitHub</p>
-        </div>
-        <TooltipProvider delay={150}>
-          <Tooltip>
-            <TooltipTrigger>
-              <Button
-                type="button"
-                variant="ghost"
-                size="icon-sm"
-                onClick={() => showConnectGitHub({})}
-                aria-label="Add GitHub account"
-              >
-                <Plus className="h-3.5 w-3.5" />
-              </Button>
-            </TooltipTrigger>
-            <TooltipContent side="top">Add GitHub account</TooltipContent>
-          </Tooltip>
-        </TooltipProvider>
+    <TooltipProvider delay={150}>
+      <div className="space-y-2">
+        {accounts.map((account) => (
+          <GitHubAccountRow
+            key={account.accountId}
+            account={account}
+            setDefaultPending={setDefaultMutation.isPending}
+            removePending={removeMutation.isPending}
+            onSetDefault={() => void setDefaultAccount(account)}
+            onRemove={() => void confirmRemove(account)}
+          />
+        ))}
       </div>
-
-      {isLoading ? (
-        <div className="text-muted-foreground flex items-center gap-2 py-3 text-sm">
-          <Loader2 className="h-4 w-4 animate-spin" />
-          Loading GitHub accounts...
-        </div>
-      ) : sortedAccounts.length === 0 ? (
-        <div className="flex items-center gap-3 rounded-lg border border-dashed border-border/70 p-3">
-          <div className="bg-muted/50 flex h-9 w-9 shrink-0 items-center justify-center rounded-md">
-            <Github className="text-muted-foreground h-4 w-4" />
-          </div>
-          <p className="text-muted-foreground min-w-0 flex-1 text-sm">
-            No GitHub accounts are connected.
-          </p>
-        </div>
-      ) : (
-        <div className="space-y-2">
-          {sortedAccounts.map((account) => (
-            <GitHubAccountRow
-              key={account.accountId}
-              account={account}
-              setDefaultPending={setDefaultMutation.isPending}
-              removePending={removeMutation.isPending}
-              onSetDefault={() => void setDefaultAccount(account)}
-              onRemove={() => void confirmRemove(account)}
-            />
-          ))}
-        </div>
-      )}
-    </section>
+    </TooltipProvider>
   );
 }
 
@@ -164,61 +174,57 @@ function GitHubAccountRow({
         </div>
         <p className="text-muted-foreground truncate text-xs">{account.host}</p>
       </div>
-      <TooltipProvider delay={150}>
-        <Tooltip>
-          <TooltipTrigger>
-            <Button
-              type="button"
-              variant="ghost"
-              size="icon-sm"
-              disabled={setDefaultPending}
-              onClick={account.isDefault ? undefined : onSetDefault}
-              aria-label={
-                account.isDefault
-                  ? `@${account.login} is the default GitHub account`
-                  : `Set @${account.login} as default GitHub account`
-              }
-            >
-              {account.isDefault ? (
-                <CircleCheck className="text-foreground" />
-              ) : (
-                <Circle className="text-foreground-muted" />
-              )}
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent side="top">
-            {account.isDefault ? 'Default account' : 'Set as default'}
-          </TooltipContent>
-        </Tooltip>
-        <Tooltip>
-          <TooltipTrigger>
-            <Button
-              type="button"
-              variant="ghost"
-              size="icon-sm"
-              disabled={removePending}
-              onClick={onRemove}
-              aria-label={`Remove @${account.login}`}
-            >
-              <X />
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent side="top">Remove account</TooltipContent>
-        </Tooltip>
-      </TooltipProvider>
+      <Tooltip>
+        <TooltipTrigger>
+          <Button
+            type="button"
+            variant="ghost"
+            size="icon-sm"
+            disabled={setDefaultPending}
+            onClick={account.isDefault ? undefined : onSetDefault}
+            aria-label={
+              account.isDefault
+                ? `@${account.login} is the default GitHub account`
+                : `Set @${account.login} as default GitHub account`
+            }
+          >
+            {account.isDefault ? (
+              <CircleCheck className="text-foreground" />
+            ) : (
+              <Circle className="text-foreground-muted" />
+            )}
+          </Button>
+        </TooltipTrigger>
+        <TooltipContent side="top">
+          {account.isDefault ? 'Default account' : 'Set as default'}
+        </TooltipContent>
+      </Tooltip>
+      <Tooltip>
+        <TooltipTrigger>
+          <Button
+            type="button"
+            variant="ghost"
+            size="icon-sm"
+            disabled={removePending}
+            onClick={onRemove}
+            aria-label={`Remove @${account.login}`}
+          >
+            <X />
+          </Button>
+        </TooltipTrigger>
+        <TooltipContent side="top">Remove account</TooltipContent>
+      </Tooltip>
     </div>
   );
 }
 
-function DefaultGitHubAccountBadge({ login }: { login: string }) {
+export function DefaultGitHubAccountBadge({ login }: { login: string }) {
   return (
-    <TooltipProvider delay={150}>
-      <Tooltip>
-        <TooltipTrigger className="inline-flex h-4.5 items-center leading-none">
-          <GitHubDefaultAccountBadge />
-        </TooltipTrigger>
-        <TooltipContent side="top">New projects will use @{login} by default.</TooltipContent>
-      </Tooltip>
-    </TooltipProvider>
+    <Tooltip>
+      <TooltipTrigger className="inline-flex h-4.5 items-center leading-none">
+        <GitHubDefaultAccountBadge />
+      </TooltipTrigger>
+      <TooltipContent side="top">New projects will use @{login} by default.</TooltipContent>
+    </Tooltip>
   );
 }
